+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 - [QA] Simplify Playground endpoint
 - [QA] Reorganize e2e tests around scopes
+- [REFACTOR] Introduce common LiveBlock component
+- [REFACTOR] Introduce common StaticBlock component
+- [FIX] Transform being applied to soon due to placeholder space character in new cells
+- [QA] Add blank `.credo.exs`
+- [REFACTOR] Make ContentEditable less hacky
+- [QA] Document ContentEditable
 
 ## 0.10.1